requirements.txt
tensorflow-cpu==2.15.0
fastapi
uvicorn
pillow
requests
